requirements.txt
pyautogui>=0.9.54
pygetwindow>=0.0.9
Pillow>=10.0.0
PyMuPDF>=1.23.0
keyboard>=0.13.5
